Delete of comments and refactor method ogin
