+++ build.gradle
@@ -11,7 +11,6 @@ sourceCompatibility = 1.8
 
 repositories {
     mavenCentral()
-    //jcenter()
 }
 
 ext {
